jardin_vigil/backend/requirements.txt
mysql-connector-python
flask
flask-cors
flask-bcrypt
requests
flask-jwt-extended
python-dotenv
flask-mail
bcrypt
jsonify